Caffe YoloV3 training patch (#7)
* Fix type issue for SSD inference detection file(.json) dump

* Add support for Yolov2 training(COCO Dataset)

* Add multi_label feature in data layer to support YoloV3

* Remove unused function

* Add YoloV3 loss layer

* Fix segmentation fault error while doing make runtest

+++ src/caffe/util/yolo_preprocess.cpp
@@ -238,8 +238,8 @@ void place_image(cv::Mat im, int w, int h, int dx, int dy, float *resized_image,
   for(c = 0; c < im.channels(); ++c){
       for(y = 0; y < h; ++y){
           for(x = 0; x < w; ++x){
-              int rx = ((float)x / w) * im.cols;
-              int ry = ((float)y / h) * im.rows;
+              float rx = ((float)x / w) * im.cols;
+              float ry = ((float)y / h) * im.rows;
               float val = bilinear_interpolate(im, rx, ry, c);
               set_pixel_with_scaling(resized_image, resize_w, resize_h,
                               im.channels(), x + dx, y + dy, c, val, scale);
